requirements.txt
Django==5.2.5
djangorestframework==3.15.2
drf_yasg==1.21.7
setuptools==70.0.0
psycopg==3.2.0
djangorestframework-simplejwt==5.4.0
django-redis==5.4.0
python-dotenv==1.1.1
sqlparse==0.5.3
asgiref==3.9.1
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.7.0
django-ratelimit==3.0.0

requests==2.32.5
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
